fix: 🐛 prevent auth files from being accidentally overwritten
Merge PR#114 to prevent auth file overwrite will cause a breaking change
such that when the user runs a command that passes both a username and
oauth2Options together to authInfo.ts then an error will be thrown.

BREAKING CHANGE: 🧨 some commands that call getCoreConnection with auth options and a
username will now throw an error.